display per message subjects in thread widget
Many senders to not quite grasp the concept of a mail thread, and this
"hijack" a thread by replying and changing the subject. In addition,
there may be good reasons for replies with changed subject, such as an
adjustment of status or the typical patch series threads with individual
topics.

Currently, all this information is hidden in collapsed view but may be
valuable for expanding the right message. Therefore, provide individual
subjects for the messages in thread view.