allow to open multiple buffers of the same type
We actively prevent opening multiple taglist or bufferlist buffers, but
so far allow to open multiple named query buffers. They cannot differ,
though, since there is nothing to configure about them.

Rather than confusing the user by behaving differently for different
buffer types, lift the restriction for all of them (but the bufferlist):
at worst there are multiple instances of buffers with the same content.

We do keep a single exception for the bufferlist buffer because that is
the one you typically use to switch to other buffers.