taglist, bufferlist: provide defaults of correct signature
The filtfun argument of these lists expects a boolean valued function of
a single argument. The current default is `lambda x: x` which in most
cases will result in no filtering, but may give the false impression
that the filtfun argument is a transformation rather than a filter.

Be explicit and specify the default no-op filter as `lambda x: True`.