taglist, bufferlist: correctly reuse buffer
Currently, we allow only one copy of these buffers and reuse them on a
second call but we fail to adjust the filter in case it changes. (The
filter is used in no call chain currently and cannot be specified from
UI.)

Adjust the commands to correctly reset the filter when the buffer is
reused.