taglist: restrict tags to displayed messages by default
Currently, the taglist command runs on the global list of tags.

Change the default behaviour to restrict the taglist to those
appearing in the displayed messages of a search or thread buffer. This
is basically the same list which `notmuch search --output=tags` would
return for the query underlying the search buffer resp. the query for
the thread displayed in a thread buffer.

Introduce a flag `--global` which allows to list all tags globally.
Binding a key to `taglist --global` restores the previous default
behaviour.