taglist: allow to restrict tags to displayed messages
Currently, the taglist command runs on the global list of tags.

Introduce a flag `msgs` which allows to restrict the taglist to those
appearing in the displayed messages of a search or thread buffer. This
is basically the same list which `notmuch search --output=tags` would
return for the query underlying the search buffer resp. the query for
the thread displayed in a thread buffer.