taglist: refine search from buffer
When a taglist buffer lists tags from a search or thead buffer (with
`--msgs`) then selecting a tag lists all messages in the db with that
tag. This feels counter-intuitive.

Instead, remember the original query and refine the query by filtering
on the tag in addition to the original query.