refactor taglist/TagWidget
taglist jumps through some hoops to amend the tags in the list by
additional information (hidden, name before translation) without urwids
column spreading kicking in, when the aim is in fact to add that text to
the tag text.

Refactor TagWidget so that callers can ask the widget to amend the
information. The default is off for callers like the search and thread
widgets.